+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @folionote/auth
 
+## 0.2.18
+
+### Patch Changes
+
+- feat: ✨ support web search tool
+- Updated dependencies
+  - @folionote/db@0.2.18
+  - @folionote/transactional@0.2.18
+
 ## 0.2.17
 
 ### Patch Changes